move out Loop into loop.py
