+++ onnx2pytorch/operations/__init__.py
@@ -11,6 +11,7 @@ from .gather import Gather
 from .gathernd import GatherND
 from .globalaveragepool import GlobalAveragePool
 from .instancenorm import InstanceNormWrapper
+from .loop import Loop
 from .lstm import LSTMWrapper
 from .matmul import MatMul
 from .nonmaxsuppression import NonMaxSuppression
@@ -48,6 +49,7 @@ __all__ = [
     "GatherND",
     "GlobalAveragePool",
     "InstanceNormWrapper",
+    "Loop",
     "LSTMWrapper",
     "MatMul",
     "NonMaxSuppression",
